requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.3.1
pymongo==4.5.0
pandas==2.1.3
numpy==1.26.2
scipy==1.11.3
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.2
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-cache2==0.2.1
pytz==2023.3
scikit-learn==1.3.2 
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0
openpyxl==3.1.2
XlsxWriter==3.1.9 